removed all references to old script directory; added exception handling for missing viral taxid

+++ generate-outputs.R
@@ -8,6 +8,7 @@ suppressMessages(library(biomartr))
 suppressMessages(library(ape))
 suppressMessages(library(taxonomizr))
 suppressMessages(library(VariantAnnotation))
+suppressMessages(library(this.path))
 
 args <- commandArgs(T)
 
@@ -19,10 +20,11 @@ e.cutoff <- as.numeric(args[5])
 ident.cutoff <- 100*(as.numeric(args[6]))
 clust.perc <- as.numeric(args[7])
 outdir <- args[8]
+script.dir <- this.dir() 
+refdir <- "/drive3/staging/agarofal/tagalong/reference/"
 
-virus.dict <- fread("~/CNV_integration_method/reference/viral_id_map_table.txt", header=F, sep="\t", stringsAsFactors = F) %>% dplyr::rename(id=1, name=2)
+virus.dict <- fread(file.path(refdir, "viral_id_map_table.txt"), header=F, sep="\t", stringsAsFactors = F) %>% dplyr::rename(id=1, name=2)
 
-refdir <- "/drive3/staging/agarofal/CNV_integration_method/reference/"
 taxa.table <- file.path(refdir, "accessionTaxa.sql")
 
 blast.frame <- tryCatch(fread(blast.path, header = F, sep = '\t', stringsAsFactors = F), error=function(e) NULL)
@@ -38,6 +40,7 @@ getmode <- function(v) {
 txdb.hg19 <- TxDb.Hsapiens.UCSC.hg19.knownGene
 
 extract.breaks <- function(clust){
+    print(clust)
     cluster.sub <- subset(blast.frame, cluster == clust)
     virus.id <- names(table(cluster.sub$sseqid))[which.max(table(cluster.sub$sseqid))]
     virus.name <- virus.dict$name[virus.dict$id == virus.id]
@@ -45,40 +48,44 @@ extract.breaks <- function(clust){
     consensus.sub <- subset(cluster.sub, sseqid == virus.id)
     virus.breakpos <- ifelse(grepl("up", clust), getmode(consensus.sub$sstart), getmode(consensus.sub$send))
     
-    gff.path <- Sys.glob(paste("/drive3/staging/agarofal/CNV_integration_method/reference/*", virus.taxid, "*gff.gz", sep=""))
-    
-    if(length(gff.path) == 0){
-      getGFF(db = "refseq", organism = virus.taxid, path=refdir)
+    if(!is.na(virus.taxid)){
       gff.path <- Sys.glob(paste(refdir, "/*", virus.taxid, "*gff.gz", sep=""))
-      rm.command <- paste("rm ", refdir, "/doc* ", refdir, "/*md5*", sep="")
-      system(rm.command)
-    }
-    
-    if(length(gff.path) != 0){
-      gff <- read_gff(gff.path)
-      gff.sub <- gff %>% filter(type == "gene", start <= virus.breakpos & end >= virus.breakpos)
-      if(nrow(gff.sub) != 0){
-        viral.gene <- paste(sapply(strsplit(sapply(strsplit(gff.sub$attribute, ";"), "[[", 3), "="), "[[", 2), collapse=";")
-      } else{
-        gene.sub <- gff %>% filter(type == "gene")
-        closest.start.sub <- gene.sub[which.min(abs(gene.sub$start - virus.breakpos)),]
-        closest.start <- closest.start.sub$start[1]
-        
-        closest.end.sub <- gene.sub[which.min(abs(gene.sub$end - virus.breakpos)),]
-        closest.end <- closest.start.sub$end[1]
-        
-        closest <- c(closest.start, closest.end)[which.min(abs(c(closest.start, closest.end) - virus.breakpos))]
-        
-        if(closest == closest.start){
-          gff.sub <- closest.start.sub
-        } else{
-          gff.sub <- closest.end.sub
-        }
+      
+      if(length(gff.path) == 0){
+        getGFF(db = "refseq", organism = virus.taxid, path=refdir)
+        gff.path <- Sys.glob(paste(refdir, "/*", virus.taxid, "*gff.gz", sep=""))
+        rm.command <- paste("rm ", refdir, "/doc* ", refdir, "/*md5*", sep="")
+        system(rm.command)
+      }
+      
+      if(length(gff.path) != 0){
+        gff <- read_gff(gff.path)
+        gff.sub <- gff %>% filter(type == "gene", start <= virus.breakpos & end >= virus.breakpos)
         if(nrow(gff.sub) != 0){
           viral.gene <- paste(sapply(strsplit(sapply(strsplit(gff.sub$attribute, ";"), "[[", 3), "="), "[[", 2), collapse=";")
         } else{
-          viral.gene <- "NA"
+          gene.sub <- gff %>% filter(type == "gene")
+          closest.start.sub <- gene.sub[which.min(abs(gene.sub$start - virus.breakpos)),]
+          closest.start <- closest.start.sub$start[1]
+          
+          closest.end.sub <- gene.sub[which.min(abs(gene.sub$end - virus.breakpos)),]
+          closest.end <- closest.start.sub$end[1]
+          
+          closest <- c(closest.start, closest.end)[which.min(abs(c(closest.start, closest.end) - virus.breakpos))]
+          
+          if(closest == closest.start){
+            gff.sub <- closest.start.sub
+          } else{
+            gff.sub <- closest.end.sub
+          }
+          if(nrow(gff.sub) != 0){
+            viral.gene <- paste(sapply(strsplit(sapply(strsplit(gff.sub$attribute, ";"), "[[", 3), "="), "[[", 2), collapse=";")
+          } else{
+            viral.gene <- "NA"
+          }
         }
+      } else{
+        viral.gene <- "NA"
       }
     } else{
       viral.gene <- "NA"
@@ -123,9 +130,9 @@ extract.breaks <- function(clust){
     write.table(cluster.human.fa, file=human.fa.path, col.names=F, row.names=F, sep="\t", quote=F)
     write.table(cluster.viral.fa, file=viral.fa.path, col.names=F, row.names=F, sep="\t", quote=F)
     
-    cdhit.command <- paste("sh /drive3/staging/agarofal/CNV_integration_method/scripts/run-cdhit.sh", sample, human.fa.path, outdir, clust.perc, sep=" ")
+    cdhit.command <- paste("sh", file.path(script.dir, "run-cdhit.sh"), sample, human.fa.path, outdir, clust.perc, sep=" ")
     system(cdhit.command)
-    cdhit.command2 <- paste("sh /drive3/staging/agarofal/CNV_integration_method/scripts/run-cdhit.sh", sample, viral.fa.path, outdir, clust.perc, sep=" ")
+    cdhit.command2 <- paste("sh", file.path(script.dir, "run-cdhit.sh"), sample, viral.fa.path, outdir, clust.perc, sep=" ")
     system(cdhit.command2)
     
     if(file.exists(human.cdhit.path)){
@@ -151,11 +158,11 @@ extract.breaks <- function(clust){
       junc.seq <- "NA"
     }
     
-    output.vec <- c(clust, human.breakchr, human.breakpos, human.type, human.gene, human.seq, virus.id, virus.name, virus.breakpos, viral.gene, viral.seq, strm, junc.seq, read.support)
+    output.vec <- c(sample, clust, human.breakchr, human.breakpos, human.type, human.gene, human.seq, virus.id, virus.name, virus.breakpos, viral.gene, viral.seq, strm, junc.seq, read.support)
     return(output.vec)
 }
 
-output.table <- do.call(rbind.data.frame, unname(lapply(unique(blast.frame$cluster), function(x) extract.breaks(x)))) %>% dplyr::rename(cluster_name=1, human_chr=2, human_breakpos=3, human_region=4, human_gene=5, human_seq=6, virus_id=7, virus_name=8, virus_breakpos=9, virus_gene=10, virus_seq=11, break_ori=12, breakpt_seq=13, frag_support=14)
+output.table <- do.call(rbind.data.frame, unname(lapply(unique(blast.frame$cluster), function(x) extract.breaks(x)))) %>% dplyr::rename(sample=1, cluster_name=2, human_chr=3, human_breakpos=4, human_region=5, human_gene=6, human_seq=7, virus_id=8, virus_name=9, virus_breakpos=10, virus_gene=11, virus_seq=12, break_ori=13, breakpt_seq=14, frag_support=15)
 
 write.table(output.table, file=paste(outdir, "/", sample, ".integration.calls.txt", sep=""), col.names=T, row.names=F, sep="\t", quote=F)
